// lib/_TC/_SeD/_V/_ST/enums/BrdBlm_TC_SeD_V_ST_eUpperCurveType.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm_TC_SeD_V_ST]
 * @author [APG] Angeli Paolo Giusto 
 * @version 0.1 APG 20231114
 * @version 0.2 APG 20231226 Estratto nel suo proprio file
 * ----------------------------------------------------------------------------
 */

/**
 * Tipologie delle curve superiori per i vari tipi di scorrimento del portone
 * sezionale a scorrimento verticale
 */
export enum BrdBlm_TC_SeD_V_ST_eUpperCurveType {

    NONE = "none",
    SO_STO = "SO_STO",
    SR_S1 = "SR_S1",
    SMB2 = "SMB2"

}